feat: Jupyter accepts children
* Fixes an issue where I forgot to allow a children prop in JupyterRoot
* Also moves the Jupyter components to a subdirectory
* Also fixes the Plasmic styling
